requirements.txt
qrcode==7.4.2
streamlit==1.29.0
pillow==9.5.0
